requirements.txt
discord.py
dotenv
pynput